feat: move component functions to MessageInteractionContext
