+++ src/structures/interfaces/messageInteraction.ts
@@ -1,5 +1,5 @@
 import { ComponentActionRow, Endpoints, InteractionResponseFlags, InteractionResponseType } from '../../constants';
-import SlashCreator from '../../creator';
+import SlashCreator, { ComponentRegisterCallback } from '../../creator';
 import { RespondFunction } from '../../server';
 import { formatAllowedMentions, FormattedAllowedMentions, MessageAllowedMentions } from '../../util';
 import Member from '../member';
@@ -300,6 +300,56 @@ class MessageInteractionContext {
     }
 
     return false;
+  }
+
+  /**
+   * Registers a component callback from the initial message.
+   * This unregisters automatically when the context expires.
+   * @param custom_id The custom ID of the component to register
+   * @param callback The callback to use on interaction
+   */
+  registerComponent(custom_id: string, callback: ComponentRegisterCallback) {
+    if (this.expired) throw new Error('This interaction has expired');
+    if (!this.initiallyResponded || this.deferred)
+      throw new Error('You must send a message before registering components');
+    if (!this.messageID)
+      throw new Error('Fetch your original message or use deferred messages before registering components');
+
+    this.creator._componentCallbacks.set(`${this.messageID}-${custom_id}`, {
+      callback,
+      expires: this.invokedAt + 1000 * 60 * 15
+    });
+  }
+
+  /**
+   * Registers a component callback from a message.
+   * This unregisters automatically when the context expires.
+   * @param message_id The message ID of the component to register
+   * @param custom_id The custom ID of the component to register
+   * @param callback The callback to use on interaction
+   */
+  registerComponentFrom(message_id: string, custom_id: string, callback: ComponentRegisterCallback) {
+    if (this.expired) throw new Error('This interaction has expired');
+    if (!this.initiallyResponded || this.deferred)
+      throw new Error('You must send a message before registering components');
+
+    this.creator._componentCallbacks.set(`${message_id}-${custom_id}`, {
+      callback,
+      expires: this.invokedAt + 1000 * 60 * 15
+    });
+  }
+
+  /**
+   * Unregisters a component callback.
+   * @param custom_id The custom ID of the component to unregister
+   * @param message_id The message ID of the component to unregister, defaults to initial message ID if any
+   */
+  unregisterComponent(custom_id: string, message_id?: string) {
+    if (!message_id) {
+      if (!this.messageID) throw new Error('The initial message ID was not provided by the context!');
+      else message_id = this.messageID;
+    }
+    return this.creator._componentCallbacks.delete(`${message_id}-${custom_id}`);
   }
 }
